requirements.txt
openai==1.35.7
python-dotenv==1.0.0
semantic-kernel==1.34.0
nest-asyncio==1.6.0